reverted 3rd-party libraries back to unmodified original versions
the Firefox AMO reviewers can't seem to make up their minds
as to what they want.

1st, the addon was rejected because of certain methods being present
in 3rd-party libraries, and I was required to sanitize them.

next, the addon was rejected because the hash/checksum
of the 3rd-party libraries doesn't match the original.

I'm honestly at a loss at this point..

3rd party libraries reverted to:
* https://github.com/chjj/marked/raw/1c188ebb55ef512653e03b0a03e60d7209a3bc73/marked.min.js
* http://cdnjs.cloudflare.com/ajax/libs/highlight.js/8.1/highlight.min.js

the `highlight.js` css themes require that the highlighted
code be wrapped in a block element having the class `hljs`.
From within the `marked.js` `highlight()` callback function,
there's no way to add a css class to the <code> element.
In previous versions, I added an additional block element
within this callback, which resulted in the DOM structure:
<pre><code><div.hljs>...highlighted code...</div>\n</code></pre>

however, the modified version of `marked.js` prevented this extra CR.
I reported its presence as an issue.
They say it's there for code readability.
I say that it's a trailing CR within a <pre> element.
But the real reason that it's more of an issue for me,
is that if the CR were appended to a series of inline elements,
then it most likely wouldn't cause an additional newline.
However, by appending it after a block element (ie: the <div>),
it's an automatic newline (stands out like a sore thumb).

by reverting back to an unmodified version of `marked.js`,
I needed to come up with a different workaround
that would allow me to wrap the highlighted code in a block
element having the css class `hljs`.
A quick look at the (unminified) source code revealed that the
option `langPrefix` could be repurposed to do exactly that.

only caveat is: the github css rules are stronger and take priority.
To work around this, all of the `highlight.js` css themes needed:
  * search  = `((?:^|,)\s*)\.hljs([\s,]|$)`
  * replace = `\1.markdown-body pre code.hljs\2`